DRY up logic handling Geo Filters.
